refactor: session 관련 테스트 분리 1. http session을 이용하는 부분을 테스트해야 하는 문제 발생  - SessionManager라는 인터페이스를 두어 실제 환경과 테스트 환경에서의 Session을 다르게 설정  - 테스트 환경 아래의 MockUserSessionManager는 MockSession을 가짐
