+++ src/main/java/com/sns/whisper/domain/user/application/session/SessionManager.java
@@ -1,0 +1,8 @@
+package com.sns.whisper.domain.user.application.session;
+
+public interface SessionManager {
+
+    void saveUser(String userId);
+
+    String extractUser(String userId);
+}
